# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Strata Saudi Public Agent Readiness API",
    "version": "2026-08-31",
    "description": "Public read-only API and discovery documentation for Strata Risk Advisory. Project inquiry submission and all contact actions require explicit user approval. Public-content policy: ai-train=no, search=yes, ai-input=yes.",
    "contact": {
      "name": "Strata Risk Advisory",
      "url": "https://www.stratasaudi.com/contact",
      "email": "advisory@stratasaudi.com"
    }
  },
...
  "servers": [
    {
      "url": "https://www.stratasaudi.com"
    }
  ],
  "paths": {
...
    "/api/a2a/message:send": {
      "post": {
        "summary": "Ask the Strata Mandate Concierge",
        "operationId": "sendA2aConciergeMessage",
        "description": "A2A v1.0 HTTP+JSON SendMessage operation. Accepts text only and answers in English or Arabic from approved public Strata resources. Raw questions and confidential project facts are not retained. Eligible client-interest activity may create only a private redacted demand signal that is never exposed through this API or used to retrain the public concierge.",
        "parameters": [
          {
            "name": "A2A-Version",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "const": "1.0"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/a2a+json": {
              "schema": {
                "$ref": "#/components/schemas/A2ASendMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A2A message response grounded in public Strata data"
          },
          "400": {
            "description": "Invalid A2A version, message, or unsupported input part"
          },
          "413": {
            "description": "Request exceeds the public-agent body limit"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "A2ASendMessageRequest": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "object",
            "required": [
              "messageId",
              "role",
              "parts"
            ],
            "properties": {
              "messageId": {
                "type": "string",
                "maxLength": 128
              },
              "contextId": {
                "type": "string",
                "maxLength": 128
              },
              "role": {
                "type": "string",
                "const": "ROLE_USER"
              },
              "parts": {
                "type": "array",
                "minItems": 1,
                "maxItems": 4,
                "items": {
                  "type": "object",
                  "required": [
                    "text"
                  ],
                  "properties": {
                    "text": {
                      "type": "string"
                    },
                    "mediaType": {
                      "type": "string",
                      "const": "text/plain"
                    }
                  },
                  "additionalProperties": false
                }
              }
            },
            "additionalProperties": false
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}